+++ src/assets/keyframes/view/keyframeview.cpp
@@ -87,7 +87,7 @@ void KeyframeView::slotAddKeyframe(int pos)
         pos = m_position;
     }
     int offset = pCore->getItemIn(m_model->getOwnerId());
-    m_model->addKeyframe(GenTime(pos + offset, pCore->getCurrentFps()), (KeyframeType) KdenliveSettings::defaultkeyframeinterp());
+    m_model->addKeyframe(GenTime(size_t(pos + offset), pCore->getCurrentFps()), (KeyframeType)KdenliveSettings::defaultkeyframeinterp());
 }
 
 void KeyframeView::slotAddRemove()
@@ -104,7 +104,7 @@ void KeyframeView::slotEditType(int type, const QPersistentModelIndex &index)
 {
     int offset = pCore->getItemIn(m_model->getOwnerId());
     if (m_model->hasKeyframe(m_position + offset)) {
-        m_model->updateKeyframeType(GenTime(m_position + offset, pCore->getCurrentFps()), type, index);
+        m_model->updateKeyframeType(GenTime(size_t(m_position + offset), pCore->getCurrentFps()), type, index);
     }
 }
 
@@ -114,7 +114,7 @@ void KeyframeView::slotRemoveKeyframe(int pos)
         pos = m_position;
     }
     int offset = pCore->getItemIn(m_model->getOwnerId());
-    m_model->removeKeyframe(GenTime(pos + offset, pCore->getCurrentFps()));
+    m_model->removeKeyframe(GenTime(size_t(pos + offset), pCore->getCurrentFps()));
 }
 
 void KeyframeView::setDuration(int dur)
@@ -133,10 +133,10 @@ void KeyframeView::slotGoToNext()
 
     bool ok;
     int offset = pCore->getItemIn(m_model->getOwnerId());
-    auto next = m_model->getNextKeyframe(GenTime(m_position + offset, pCore->getCurrentFps()), &ok);
+    auto next = m_model->getNextKeyframe(GenTime(size_t(m_position + offset), pCore->getCurrentFps()), &ok);
 
     if (ok) {
-        emit seekToPos(qMin(next.first.frames(pCore->getCurrentFps()) - offset, m_duration - 1));
+        emit seekToPos(qMin((int)next.first.frames(pCore->getCurrentFps()) - offset, m_duration - 1));
     } else {
         // no keyframe after current position
         emit seekToPos(m_duration - 1);
@@ -154,7 +154,7 @@ void KeyframeView::slotGoToPrev()
     auto prev = m_model->getPrevKeyframe(GenTime(m_position + offset, pCore->getCurrentFps()), &ok);
 
     if (ok) {
-        emit seekToPos(qMax(0, prev.first.frames(pCore->getCurrentFps()) - offset));
+        emit seekToPos(qMax(0, (int)prev.first.frames(pCore->getCurrentFps()) - offset));
     } else {
         // no keyframe after current position
         emit seekToPos(m_duration);
@@ -255,7 +255,7 @@ void KeyframeView::mouseDoubleClickEvent(QMouseEvent *event)
         }
 
         // add new keyframe
-        m_model->addKeyframe(position, (KeyframeType) KdenliveSettings::defaultkeyframeinterp());
+        m_model->addKeyframe(position, (KeyframeType)KdenliveSettings::defaultkeyframeinterp());
     } else {
         QWidget::mouseDoubleClickEvent(event);
     }
