+++ src/scopes/colorscopes/rgbparade.cpp
@@ -117,7 +117,7 @@ QImage RGBParade::renderHUD(uint)
         } else if (valY > scopeRect().height() - bottom) {
             valY = scopeRect().height() - bottom;
         }
-        int val = 255 * (1 - ((float)y / (scopeRect().height() - RGBParadeGenerator::distBottom)));
+        int val = 255 * (int)(1. - ((float)y / (float)(scopeRect().height() - RGBParadeGenerator::distBottom)));
         davinci.drawText(x, valY, QVariant(val).toString());
     }
 
@@ -133,7 +133,7 @@ QImage RGBParade::renderGfxScope(uint accelerationFactor, const QImage &qimage)
     int paintmode = ui->paintMode->itemData(ui->paintMode->currentIndex()).toInt();
     QImage parade = m_rgbParadeGenerator->calculateRGBParade(m_scopeRect.size(), qimage, (RGBParadeGenerator::PaintMode)paintmode, m_aAxis->isChecked(),
                                                              m_aGradRef->isChecked(), accelerationFactor);
-    emit signalScopeRenderingFinished(start.elapsed(), accelerationFactor);
+    emit signalScopeRenderingFinished((uint)start.elapsed(), accelerationFactor);
     return parade;
 }
